+++ modules/paras_selection/paras_selection.py
@@ -15,24 +15,9 @@ if __name__ == '__main__':
     ## to train ParasSelection model then
     logging.info("1. Create golden paras.")
 
+    GoldenParas().generate_goldenParas()
 
-    ## Check whether total files are created.
-    ## Since train, valid and test dataset cannot fit into memory,
-    ## I read part of each and save to individual file.
-    ## Therefore, assume I split each dataset into 8 (args.n_shards) parts,
-    ## the total number of files creates are 3 * 8 = 24.
-    n_files = 0
-    for ith in range(3 * args.n_shards):
-        if check_file_existence(PATH['golden_paras'].replace("[N_PART]", str(ith))):
-            n_files += 1
-        else:
-            break
-
-    if n_files < 3 * args.n_shards:
-        GoldenParas().generate_goldenParas(n_files)
-    else:
-        logging.info("=> Golden paras files are created sufficiently.")
-
+    logging.info("=> Data files are created and saved sufficiently.")
 
 
     ########################