+++ modules/paras_selection/utils.py
@@ -3,6 +3,8 @@ import json
 import re
 import os
 
+from datasets.arrow_dataset import Dataset
+from datasets.load import load_dataset
 from transformers import BertTokenizer
 import pandas as pd
 import spacy
@@ -73,25 +75,23 @@ class GoldenParas():
 
         return entities
 
-    # def f_kernel(data: list, queue):
-    #     for data_point in data:
-    #         queue.put(PreprocessingHelper(data_point).preprocessed_datapoint)
-    # def f_findGolden(self, row: list) -> dict:
+
     def f_findGolden(self, data: list, queue):
         """Function later used in Pool of multiprocessing. This function export entities
         and find golden passage for each paragraphs. Bert Tokenizer is used also.
 
         Args:
-            row (list): row of data
+            document (list): document of data
 
         Returns:
             dict: dict containing question and golden paras.
         """
-        for row in data:
-            id_doc      = row['document']['id']
-            question    = row['question']['text']
-            answer1     = row['answers'][0]['text']
-            answer2     = row['answers'][1]['text']
+        for document in data:
+            id_doc      = document['document']['id']
+            paragraphs  = document['document']['text']
+            question    = document['question']['text']
+            answer1     = document['answers'][0]['text']
+            answer2     = document['answers'][1]['text']
 
             ##################
             ### Get entities from question
@@ -109,7 +109,7 @@ class GoldenParas():
             ### List contains goldeness of paragraphs
             goldeness               = list()
 
-            for paragraph in row['document']['text']:
+            for paragraph in paragraphs:
                 #### Determine goldeness of paragraph
                 n_found_entities    = 0
 
@@ -130,21 +130,25 @@ class GoldenParas():
                 para_tokenized  = self.BERT_tokenizer.tokenize(paragraph)
                 tokenized_paragraphs.append(self.BERT_tokenizer.convert_tokens_to_ids(para_tokenized))
 
+            document['document']['text']         = para_tokenized
+            document['document']['goldeness']    = goldeness
 
+            
 
             ##################
             ### Return a dict
             ##################
-            queue.put({
-                'id_document'    : id_doc,
-                "question"       : self.BERT_tokenizer.convert_tokens_to_ids(question.split(' ')),
-                "question_plain" : question,                ## This field's existence is for keep tracking in later step
-                "goldeness"      : goldeness,
-                "paragraphs"     : tokenized_paragraphs
-            })
+            # queue.put({
+            #     'id_document'    : id_doc,
+            #     "question"       : self.BERT_tokenizer.convert_tokens_to_ids(question.split(' ')),
+            #     "question_plain" : question,                ## This field's existence is for keep tracking in later step
+            #     "goldeness"      : goldeness,
+            #     "paragraphs"     : tokenized_paragraphs
+            # })
+            queue.put(document)
 
 
-    def generate_goldenParas(self, i: int) -> list:
+    def generate_goldenParas(self) -> list:
         """Read paragraphs, use keyword method to determine
         golden passages for each question. Additionally, BertTokenizer
         is applied.
@@ -157,15 +161,110 @@ class GoldenParas():
         """
         ## For each document, use keyword technique
         ## to determine golden passage
-        for nth, path in enumerate(self.paths_rawParas[i:]):
-            logging.info(f"= Process file {nth+i}: {os.path.split(path)[1]}")
+        for path in self.paths_rawParas:
+            split, n_shard  = re.findall(r"\_(train|test|valid)\_(\d+)\.pkl", path)[0] 
+            path_storeDat   = PATH['processed_data'].replace("[N_SHARD]", n_shard).replace("[SPLIT]", split)
 
-            ### Load dataset from path
-            with open(path, 'r') as dat_file:
-                dataset = json.load(dat_file)
+            ### Check whether this file is processed (folder specified by 'path' existed)
+            if os.path.isdir(path_storeDat):
+                continue
 
-            ### Load dataset from path
-            list_golden_paras   = ParallelHelper(self.f_findGolden, dataset,
-                                                 args.num_proc).launch()
 
-            save_object(PATH['golden_paras'].replace("[N_PART]", str(nth)), list_golden_paras)
+            logging.info(f"= Process file: {os.path.split(path)[1]}")
+
+            ## NOTE: This line is commented for testing
+            # ### Load shard from path
+            # with open(path, 'r') as dat_file:
+            #     dataset = json.load(dat_file)
+
+            # ### Process shard of dataset
+            # list_golden_paras   = ParallelHelper(self.f_findGolden, dataset,
+            #                                      args.num_proc).launch()
+
+            dataset = load_dataset("pandas", data_files=path)
+
+            dataset = dataset.map(self.f_findGolden2, num_proc=args.num_proc,
+                        # remove_columns=['kind', 'url', 'file_size', 'word_count', 'start',
+                        #                 'end', 'summary'])
+                        remove_columns=['document', 'question', 'answers'])
+
+            
+            # dataset.save_to_disk(path_storeDat)
+            save_object(path_storeDat, pd.DataFrame(dataset), is_dataframe=True)
+
+
+
+
+    def f_findGolden2(self, document: dict):
+        """Function later used in Pool of multiprocessing. This function export entities
+        and find golden passage for each paragraphs. Bert Tokenizer is used also.
+
+        Args:
+            document (list): document of data
+
+        Returns:
+            dict: dict containing question and golden paras.
+        """
+
+        paragraphs  = document['document']['text']
+        question    = document['question']['text']
+        answer1     = document['answers'][0]['text']
+        answer2     = document['answers'][1]['text']
+
+        ##################
+        ### Get entities from question
+        ### and answers
+        ##################
+        entities            = self.extract_entities(question, answer1, answer2)
+
+
+        ##################
+        ### Start finding golden passages
+        ##################
+
+        ### This list contains tokenized and converted-to-ids paragraphs
+        tokenized_paragraphs    = list()
+        ### List contains goldeness of paragraphs
+        goldeness               = list()
+
+        for paragraph in paragraphs:
+            #### Determine goldeness of paragraph
+            n_found_entities    = 0
+
+            for ent in entities:
+                try:
+                    if re.search(ent, paragraph):
+                    # if paragraph.find(ent):
+                        n_found_entities    += 1
+                except re.error:
+                    pass
+
+            if n_found_entities > 1:
+                goldeness.append(1)
+            else:
+                goldeness.append(0)
+
+            #### Tokenize and convert to id foreach paragraph
+            para_tokenized  = self.BERT_tokenizer.tokenize(paragraph)
+            tokenized_paragraphs.append(self.BERT_tokenizer.convert_tokens_to_ids(para_tokenized))
+
+
+        ##################
+        ### Update fields in 'document'
+        ##################
+        document['doc_id']          = document['document']['id']
+        document['doc_text']        = para_tokenized
+        document['question_text']   = document['question']['text']
+        document['question_tokens'] = self.BERT_tokenizer.convert_tokens_to_ids(document['question']['tokens'])
+        document['answer1_text']    = document['answers'][0]['text']
+        document['answer1_tokens']  = self.BERT_tokenizer.convert_tokens_to_ids(document['answers'][0]['tokens'])
+        document['answer2_text']    = document['answers'][1]['text']
+        document['answer2_tokens']  = self.BERT_tokenizer.convert_tokens_to_ids(document['answers'][1]['tokens'])
+        # document['document']['text']        = para_tokenized
+        # document['document']['goldeness']   = goldeness
+        # document['question']['tokens']      = self.BERT_tokenizer.convert_tokens_to_ids(document['question']['tokens'])
+        # document['answers'][0]['tokens']    = self.BERT_tokenizer.convert_tokens_to_ids(document['answers'][0]['tokens'])
+        # document['answers'][1]['tokens']    = self.BERT_tokenizer.convert_tokens_to_ids(document['answers'][1]['tokens'])
+
+
+        return document